codigos de error

+++ src/controllers/tarjetasBancaria.controller.js
@@ -3,24 +3,30 @@ import {TarjetaBancaria} from '../models/TarjetaBancaria.js';
 export const createTarjetaBancaria = async (req, res) =>{
     try {
         const {numero, fechaVencimiento, idUsuario} = req.body;
-        const newTarjetaBancaria = await TarjetaBancaria.create({
-          numero,
-          fechaVencimiento,
-          idUsuario
-        });
-    res.status(201).json(newTarjetaBancaria)
+        if (idUsuario>0) {
+            const newTarjetaBancaria = await TarjetaBancaria.create({
+                numero,
+                fechaVencimiento,
+                idUsuario
+            });
+            res.status(201).json(newTarjetaBancaria)
+        } else {
+            res.status(400).json({message: "Usuario invalido"});
+        }
+        
     } catch (error) {
-        res.json({error})
+        res.json({error:`${error}`})
+        console.log(error)
     }
 }
 
 export const getTarjetasBancarias = async (req, res) =>{
     try{
-        let tarjetasBancarias = await TarjetaBancaria.findAll();
-        if(tarjetasBancarias == null){
-            res.status(204).json();
-        }else{
+        const tarjetasBancarias = await TarjetaBancaria.findAll();
+        if(tarjetasBancarias!= null){
             res.json(tarjetasBancarias)
+        }else{
+            res.status(204).json();
         }
     } catch (error){
         res.json({error})
@@ -32,32 +38,39 @@ export const getTarjetaBancaria = async(req, res) =>{
     const {numero} = req.params;
 
     try {
-        const tarjetaB = await TarjetaBancaria.findOne({
+        const tarjetaBancaria = await TarjetaBancaria.findOne({
             where: {
                 numero,
             }
         });
-        if(tarjetaB == null){
-            res.status(204).json();
-        }else{
+        if(tarjetaBancaria != null){
             res.json(numero)
+        }else{
+            res.status(204).json();
         }
     } catch (error) {
         console.log(error);
-        res.json({error:"No hay ninguna tarjeta bancaria registrada con ese numero"})
+        res.json({error:`${error}`})
     }    
 }
 
 export const deleteTarjetaBancaria = async(req,res) => {
     try {
         const {numero} = req.params;
-        await TarjetaBancaria.destroy({
-            where: {
-                numero,
-            }
-        });
-        return res.sendStatus(204);
+        const tarjetaBancaria = await tarjetaBancaria.findByPk(numero);
+        if (tarjetaBancaria!=null) {
+            await TarjetaBancaria.destroy({
+                where: {
+                    numero,
+                }
+            });
+            return res.sendStatus(204);
+        } else {
+            return res.sendStatus(204);
+        }
+        
     } catch (error) {
-        res.json({error:"No hay ninguna tarjeta bancaria registrada para borrar con ese id"})
+        console.log(error)
+        res.json({error:`${error}`})    
     }
 }
